requirements.txt
python-dotenv
requests
selenium
roman
inquirer
prettytable